refactor: Phase 3a — move pattern compilation into formatparse-core
Port parse_pattern, parse_field, parse_format_spec, and helpers to
formatparse-core with FormatParseError (no PyO3). PyO3 keeps a thin
pattern module that re-exports the core API and maps narrow pattern
errors to PatternParseMismatch for parse/compile compatibility.
